requirements.txt
click==6.7
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
mysqlclient==1.3.12
PyJWT==1.5.3
python-dotenv==0.7.1
pytz==2017.3
whitenoise==3.3.1
